Implement burn state command
