+++ packages/cli/src/commands/archive.ts
@@ -31,7 +31,7 @@ export function formatArchiveStatusLines(status: ArchiveStatus): string[] {
   const lines: string[] = [];
   lines.push(`archive: ${status.archivePath}`);
   if (!status.exists) {
-    lines.push('  status: not built yet - run `burn rebuild archive`');
+    lines.push('  status: not built yet - run `burn state rebuild archive`');
     return lines;
   }
   lines.push(`  schema version: ${status.archiveVersion}`);
